Add CLI command openvidu:room-create, refactor other openvidu commands
